refactor: change name of the rules
