+++ src/index.ts
@@ -1,5 +1,5 @@
-import i18nOnlyIdentifiersRule from './rules/i18n-only-identifiers'
-import missingLinguiTransformationRule from './rules/missing-lingui-transformation'
+import noExpressionInMessageRule from './rules/no-expression-in-message'
+import noUnlocalizedStringsRule from './rules/no-unlocalized-strings'
 import noSingleTagToTranslateRule from './rules/no-single-tag-to-translate'
 import noSingleVariablesToTranslateRule from './rules/no-single-variables-to-translate'
 import tCallInFunctionRule from './rules/t-call-in-function'
@@ -7,8 +7,8 @@ import textRestrictionsRule from './rules/text-restrictions'
 import noTransInsideTransRule from './rules/no-trans-inside-trans'
 
 export const rules = {
-  'i18n-only-identifiers': i18nOnlyIdentifiersRule,
-  'missing-lingui-transformation': missingLinguiTransformationRule,
+  'no-expression-in-message': noExpressionInMessageRule,
+  'no-unlocalized-strings': noUnlocalizedStringsRule,
   'no-single-tag-to-translate': noSingleTagToTranslateRule,
   'no-single-variables-to-translate': noSingleVariablesToTranslateRule,
   't-call-in-function': tCallInFunctionRule,
